Pause refreshing data stream when user logged out, try to reload data when during started fetch getting error, correct logout
Actually, it would be best to provide the RatesBloc in the Root Screen, but in this case with redirects, we can afford it so as not to download a large amount of data after exiting/logging in. The disadvantage of this approach is to remember, if anything, that the rates update stream must be stopped and started again. In a good way, it would be possible to create an authorization service whose functions would be associated with the Rates block. And it would be even better to just update the bloc every time when user login/logout.